Processor API

+++ midi/tuning.go
@@ -5,7 +5,7 @@ import "math"
 // Tuning is an interface for a chromatic scale.
 type Tuning interface {
 	// A4Hz returns the frequency of the note A4.
-	A4Hz() float64
+	A4Hz() float32
 }
 
 // StdTuning is an A440 tuning.
@@ -13,7 +13,7 @@ var StdTuning = stdTuning{}
 
 type stdTuning struct{}
 
-func (stdTuning) A4Hz() float64 {
+func (stdTuning) A4Hz() float32 {
 	return 440
 }
 
@@ -29,18 +29,18 @@ func clampKey(k int) uint8 {
 
 // Key returns the MIDI key value from tuning t and frequency f.
 // A4 is MIDI key 69 for instance.
-func Key(t Tuning, f float64) int {
-	return int(69 + 12*math.Log2(f/t.A4Hz()))
+func Key(t Tuning, f float32) int {
+	return int(69 + 12*math.Log2(float64(f)/float64(t.A4Hz())))
 }
 
 // Pitch returns the pitch of the midi key in tuning t.
-func Pitch(t Tuning, key int) float64 {
-	return Tone(t, float64(key))
+func Pitch(t Tuning, key int) float32 {
+	return Tone(t, float32(key))
 }
 
 // Tone returns the tone of the fractional midi key in tuning t.
-func Tone(t Tuning, key float64) float64 {
-	return t.A4Hz() * math.Pow(2, (key-69)/12)
+func Tone(t Tuning, key float32) float32 {
+	return t.A4Hz() * float32(math.Pow(2, (float64(key)-69)/12))
 }
 
 // Note constants starting at octave 0.
